fix: select certain fields from User table instead of all fields
Previously the functions getFollowing and getFollowers from the
followCRUD module would select all fields from the User table for
followers or followed users. This included the password field, and for
security purposes that field is no longer selected. This change is
reflected in the tests for the db.test.js suite, and the example users
in the exampleUsers.js module have been changed to reflect this as well.