fix: sign jwt with just user_id instead of entire user obj, other fixes
In a previous commit, I had refactored the login authenticate
functionality to set the jwt in an Authorization header in the response,
in an effort to be more secure. However upon further reading/research
I learned that the Authorization header is a request header, not a
response header, and that it is ok to respond with the jwt in json.
Instead it is the client that sends the jwt in an Authorization header
in following requests, which will be implemented on the frontend.